Pass C extension upcall pointer arguments as plain long addresses
* P, F and A arguments were converted to a fresh Truffle::FFI::Pointer
  (or NativeValueArray) per upcall, which the Ruby side then mostly
  converted right back to an address with Primitive.interop_as_pointer
  or passed to Primitive.cext_invoke_* (which takes raw addresses).
  Pass the raw long address instead, so no allocation and no conversion
  is needed (NULL is 0 instead of nil).
* With that F (function pointer) and A (VALUE array) arguments behave
  exactly like P, so consolidate the three carriers into P.

+++ src/main/c/cext/encoding.c
@@ -368,7 +368,7 @@ VALUE rb_enc_str_new(const char *ptr, long len, rb_encoding *enc) {
   if (ptr == NULL) {
     return rb_tr_up_send1_force_encoding(rb_tr_up_rb_str_new_nul(len), rb_enc_from_encoding(enc));
   }
-  return rb_tr_up_rb_enc_str_new_native(ptr, len, (long) enc);
+  return rb_tr_up_rb_enc_str_new_native(ptr, len, enc);
 }
 
 #undef rb_enc_str_new_cstr
